backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
SQLAlchemy==2.0.23
psycopg2==2.9.9
alembic==1.13.2
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.6
PyJWT==2.8.0
bcrypt==3.2.0
passlib[bcrypt]==1.7.4
pydantic==2.7.4